fix(install.sh): test if init-script exists before copy

+++ install.sh
@@ -14,7 +14,11 @@ hash git >/dev/null && /usr/bin/env git clone git://github.com/edouard-lopez/ssh
   echo "git not installed"
   exit
 }
+initDir=/etc/init.d
 
 # copy to System-V style init script directory
 cd sshuttle-helper
-sudo cp ./proxy-ssh /etc/init.d/
+if [ -e "$initDir/proxy-ssh" ]; then
+  echo "Adding init script..."
+  sudo cp ./proxy-ssh /etc/init.d/
+fi
